Renamed HelloApi and HelloGui, come on guys, that was supposed to be the first thing you do...
