+++ api/src/main/java/api/Api.java
@@ -1,4 +1,4 @@
-package helloApi;
+package api;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -9,7 +9,7 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.ArrayList;
-public class HelloApi {
+public class Api {
     public static void helloApi() {
         System.out.println("Hello from api");
