+++ api/src/main/java/module-info.java
@@ -2,7 +2,7 @@ module MTG.Deck.Builder.api {
     requires org.json;
     requires MTG.Deck.Builder.shared;
     //requires animal.sniffer.annotations;
-    exports helloApi;
+    exports api;
 
 
 }
