+++ gui/src/main/java/gui/GUI.java
@@ -1,4 +1,4 @@
-package helloGui;
+package gui;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
